server/requirements.txt
fastapi==0.111.0
python-dotenv>=1.0.0
uvicorn[standard]==0.30.1
websockets==12.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt<4.0.0
pyotp==2.9.0
qrcode[pil]==7.4.2
Pillow>=11.0.0
aiosqlite==0.20.0
